app/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
Markdown==3.6
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
uuid==1.30
django-ratelimit==4.1.0
